requirements.txt
# requirements for trend_scanner (Project Aegis)
# Generated by GitHub Copilot. Review and pin versions as needed.  
# Note: Some packages may require specific versions for compatibility.
# If a package can't be found on PyPI, install it from the project's private index or source distribution.

# CrewAI for agent orchestration
crewai
crewai-tools

# Google Gemini LLM backend
google-generativeai
langchain-google-genai

# Web scraping and content extraction
beautifulsoup4
newspaper3k
readability-lxml
trafilatura

# Reddit API
praw

# Optional helpers
dotenv

# Core HTTP / parsing
requests
feedparser

# Databases
pymongo

# News APIs
newsapi-python

# ML and data processing
scikit-learn